+++ src/Stormancer.Plugins/GameFinder/Stormancer.Server.Plugins.GameFinder/Models/GameFinderResult.cs
@@ -36,7 +36,7 @@ namespace Stormancer.Server.Plugins.GameFinder
         public List<Game> Games { get; } = new List<Game>();
 
         /// <summary>
-        /// A list of tickets to open game sessions, filled by <see cref="IGameFinder"/>.
+        /// A list of tickets to open game sessions, filled by <see cref="IGameFinderAlgorithm"/>.
         /// </summary>
         /// <seealso cref="OpenGameSession"/>
         public List<OpenGameSessionTicket> GameSessionTickets { get; } = new List<OpenGameSessionTicket>();